+++ internal/server/model_request_handler.go
@@ -70,7 +70,7 @@ type requestGroup struct {
 // Query parameters:
 //   - limit: maximum number of requests to return (default: 100, max: 1000)
 //   - scenario / provider / status: optional exact-match filters
-func (h *WebUIHandler) GetModelRequests(c *gin.Context) {
+func (h *WebHandler) GetModelRequests(c *gin.Context) {
 	if h.deps.MultiLogger == nil {
 		c.JSON(http.StatusServiceUnavailable, gin.H{"error": "Logger not available"})
 		return
@@ -119,7 +119,7 @@ func (h *WebUIHandler) GetModelRequests(c *gin.Context) {
 }
 
 // GetModelRequestDetail returns the full event timeline for a single request id.
-func (h *WebUIHandler) GetModelRequestDetail(c *gin.Context) {
+func (h *WebHandler) GetModelRequestDetail(c *gin.Context) {
 	if h.deps.MultiLogger == nil {
 		c.JSON(http.StatusServiceUnavailable, gin.H{"error": "Logger not available"})
 		return
@@ -154,7 +154,7 @@ func (h *WebUIHandler) GetModelRequestDetail(c *gin.Context) {
 
 // collectRequestGroups scans the HTTP, model_request and smart_routing memory
 // sinks and groups every entry by its request_id.
-func (h *WebUIHandler) collectRequestGroups() map[string]*requestGroup {
+func (h *WebHandler) collectRequestGroups() map[string]*requestGroup {
 	groups := make(map[string]*requestGroup)
 
 	ingest := func(source obs.LogSource) {
